requirements.txt
torch
torchvision
matplotlib
numpy
pandas
scikit-learn
tqdm
Pillow
seaborn
PyYAML
